flights/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
Flask-Cors==5.0.0
opentelemetry-distro[otlp]
opentelemetry-instrumentation-logging